code-plugin/pyproject.toml
[tool.poetry]
name = "snowbrain"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
modal-client = "^0.49.2130"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
numpy = "^1.24.3"
matplotlib = "^3.7.1"
pillow = "^9.5.0"
seaborn = "^0.12.2"
snowflake-connector-python = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

